+++ tests/testthat/test-data_arrange.R
@@ -1,4 +1,4 @@
-library(dplyr)
+library(poorman)
 df <- head(mtcars)
 df$character <- c("a", "b", "b", "c", "c", "a")
 
@@ -44,10 +44,17 @@ test_that("data_arrange works without columns", {
 })
 
 test_that("data_arrange ignores wrong names if safe = TRUE", {
-  expect_equal(data_arrange(df, "foo"), df)
-  expect_equal(
-    data_arrange(df, c("gear", "foo")),
-    data_arrange(df, "gear")
+  expect_warning(
+    expect_equal(data_arrange(df, "foo"), df),
+    regexp = "don't exist"
+  )
+
+  expect_warning(
+    expect_equal(
+      data_arrange(df, c("gear", "foo")),
+      data_arrange(df, "gear")
+    ),
+    regexp = "don't exist"
   )
 })
 
@@ -61,5 +68,29 @@ test_that("data_arrange errors if not coercable to data frame", {
     data_arrange(list(a = 1:5, b = letters[5:1]), select = "b"),
     structure(list(a = 5:1, b = c("a", "b", "c", "d", "e")), row.names = 5:1, class = "data.frame"),
     ignore_attr = TRUE
+  )
+})
+
+test_that("data_arrange works with grouped df", {
+  set.seed(123)
+  x <- mtcars[sample(1:nrow(mtcars), 10, replace = TRUE), c("cyl", "mpg", "drat")]
+  g <- data_group(x, cyl)
+
+  expected <- data.frame(
+    cyl = c(4, 4, 4, 6, 6, 8, 8, 8, 8, 8),
+    mpg = c(22.8, 30.4, 32.4, 17.8, 19.2, 10.4, 15, 15.2, 15.5, 18.7),
+    drat = c(3.85, 4.93, 4.08, 3.92, 3.92, 2.93, 3.54, 3.07, 2.76, 3.15)
+  )
+  class(expected) <- c("grouped_df", "data.frame")
+  rownames(expected) <- c(
+    "Datsun 710", "Honda Civic", "Fiat 128", "Merc 280C", "Merc 280",
+    "Cadillac Fleetwood", "Maserati Bora", "Merc 450SLC", "Dodge Challenger",
+    "Hornet Sportabout"
+  )
+  attributes(expected)$groups <- attributes(g)$groups
+
+  expect_equal(
+    data_arrange(g, "mpg"),
+    expected
   )
 })